% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeProb.R
\name{computeProb}
\alias{computeProb}
\title{Compute Probability Vector Elements}
\usage{
computeProb(DWTmat, Qvec, useBFIC = FALSE, isDataVector = FALSE)
}
\arguments{
\item{DWTmat}{The matrix of discrete wavelet transform of data values.}

\item{Qvec}{The discrete wavelet transform of the idealized time series with shift.}

\item{useBFIC}{set to TRUE to compute Bayes factor information content.}

\item{isDataVector}{set to TRUE for 1-d time series}
}
\description{
This is a helper function for changePointDetect.
}
