% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createTimeSeries.R
\name{createTimeSeries}
\alias{createTimeSeries}
\title{Create Time Series}
\usage{
createTimeSeries(mu1, mu2, sigma, n = 128, tau = 72)
}
\arguments{
\item{mu1}{A vector of length dimension giving mean before change point, or scalar if constant across dimensions.}

\item{mu2}{A vector of length dimesnion giving mean after change point, or scalar if constant across dimensions.}

\item{sigma}{The covariance matrix used for generation. Defaults to the identity.}

\item{n}{The length of the time series to create.}

\item{tau}{The location of the change point.}
}
\value{
A vector of length n containing multivariate normal entries.
}
\description{
This function creates a multi-dimensional nomral time series with normal entries and a possible change point in the mean. Running with no
arguments creates five dimensional normal time series with mean 0, mean 1 and identity covariance matrix.
}
\examples{
createTimeSeries()
}
