% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotChangePoint.R
\name{plotChangePoint}
\alias{plotChangePoint}
\title{Change point plotter}
\usage{
plotChangePoint(
  M1,
  M2,
  pad1,
  probvec,
  wid,
  m,
  n,
  data,
  indices,
  showall = FALSE
)
}
\arguments{
\item{M1}{Max of probability vector [calculated in changePointDetect]}

\item{M2}{If M1 > 3, good evidence of change point [calculated in changePointDetect]}

\item{pad1}{padding  [calculated in changePointDetect]}

\item{probvec}{idealized data set and its discrete wavelet transform [calculated in changePointDetect]}

\item{wid}{number of columns of data (width) [calculated in changePointDetect]}

\item{m}{row length [calculated in changePointDetect]}

\item{n}{length [calculated in changePointDetect]}

\item{data}{change point dataset [calculated in changePointDetect]}

\item{indices}{indices of change point [calculated in changePointDetect]}

\item{showall}{set to TRUE to see the top three candidate plots based on highest BFIC value}
}
\description{
Plots change points. Shows best plot or the top three plots based on BFIC value.
}
