% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SVDetectChangePoint.R
\name{SVDetectChangePoint}
\alias{SVDetectChangePoint}
\title{Singular Value of the Laplacian Change Point Detection}
\usage{
SVDetectChangePoint(
  multiSeries,
  useGaussian = FALSE,
  useBFIC = TRUE,
  showplot = TRUE
)
}
\arguments{
\item{multiSeries}{The high dimensional time series. Should be a matrix, with each row being an observation of the time series.}

\item{useGaussian}{Set to TRUE if you want to use a random Gaussian projection. Default is random matrix of +- 1.}

\item{useBFIC}{Optional argument to use BFIC to decide change point location.}

\item{showplot}{set to TRUE to see plot of 1-d time series and probability plot.}

\item{setdetail}{Optional argument to set the detail level you wish to use. Default is all details.}
}
\description{
Computes a random projection with Bernoulli or Gaussian entries of each element in a time series, then calls detectChangePoint to determine where the
change point occurs. In most cases of real data, the BFIC will be significantly greater than 3; it is an open
question to determine good values of the BFIC for various types of problems.
}
